refactor(Lib): Refactor static color lib for add multiple libs
